requirements.txt
asgiref==3.6.0
astroid==2.13.2
autopep8==2.0.1
dill==0.3.6
dj-database-url==1.2.0
Django==4.1.5
djangorestframework==3.14.0
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==2.6.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pylint==2.15.10
pytz==2022.7
sqlparse==0.4.3
tomlkit==0.11.6
typing_extensions==4.4.0
wrapt==1.14.1
